refactor: move and rename UpdatePanelGit to Git
